moved chunk_file up one level
