+++ redbox/parsing/__init__.py
@@ -1,1 +1,5 @@
 """Redbox file parsing module."""
+
+from redbox.parsing.file_chunker import chunk_file
+
+__all__ = ["chunk_file"]
